# [PERF]: better locking of uncommitted tracking maps (decrease compaction time by 3x)