## Description of changes

Instead of acquiring 2 locks for every token, we now acquire the locks at the document level. This improves compaction time by up to 3x.

I ran an experiment where I:

1. Added 15,000 records
2. Compacted the records
3. Added 9,600 records
4. Compacted the records

The partition size was 10,000.

Results (also includes changes from https://github.com/chroma-core/chroma/pull/2729 up to https://github.com/chroma-core/chroma/pull/2729/commits/14a744bce2b70fcf458de1adb68c203fd739c174):

**Before**:

- First compaction: 110s
- Second compaction: 54s

**After**:

- First compaction: 33s
- Second compaction: 40s

Here's a trace file for the two runs: [locking-changes.trace.zip](https://github.com/user-attachments/files/16790134/locking-changes.trace.zip).

(Run 1 is after these changes, run 2 is before these changes.)

## Test plan
*How are these changes tested?*

covered by existing tests

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need to make documentation changes in the [docs repository](https://github.com/chroma-core/docs)?*

n/a